+++ core/api/openapi/src/main/kotlin/org/http4k/contract/Tag.kt
@@ -1,3 +1,9 @@
 package org.http4k.contract
 
-data class Tag(val name: String, val description: String? = null)
+data class Tag(
+    val name: String,
+    val description: String? = null,
+    val summary: String? = null,
+    val parent: String? = null,
+    val kind: String? = null
+)