+++ core/api/openapi/src/main/kotlin/org/http4k/contract/openapi/v3/OpenApi3.kt
@@ -115,7 +115,8 @@ class OpenApi3<NODE : Any>(
                 ),
                 servers,
                 webhookPaths,
-                version.toString()
+                version.toString(),
+                version.jsonSchemaDialect
             )
         )