+++ core/api/openapi/src/test/kotlin/org/http4k/contract/openapi/v3/OpenApi3AutoTest.kt
@@ -6,6 +6,7 @@ import org.http4k.contract.AutoContractRendererContract
 import org.http4k.contract.jsonschema.v3.FieldMetadata
 import org.http4k.contract.openapi.AddSimpleFieldToRootNode
 import org.http4k.contract.openapi.ApiInfo
+import org.http4k.contract.openapi.ApiLicense
 import org.http4k.contract.openapi.OpenAPIJackson
 import org.http4k.core.Uri
 import kotlin.reflect.full.createType
@@ -13,7 +14,7 @@ import kotlin.reflect.full.createType
 class OpenApi3AutoTest : AutoContractRendererContract<JsonNode>(
     OpenAPIJackson,
     OpenApi3(
-        ApiInfo("title", "1.2", "module description"),
+        ApiInfo("title", "1.2", "module description", "a short summary", ApiLicense.Apache2_0),
         OpenAPIJackson,
         listOf(AddSimpleFieldToRootNode),
         listOf(ApiServer(Uri.of("http://localhost:8000"), "a very simple API")),