+++ core/api/openapi/src/test/kotlin/org/http4k/contract/openapi/v3/OpenApi3Test.kt
@@ -4,11 +4,12 @@ import argo.jdom.JsonNode
 import org.http4k.contract.ContractRendererContract
 import org.http4k.contract.openapi.AddSimpleFieldToRootNode
 import org.http4k.contract.openapi.ApiInfo
+import org.http4k.contract.openapi.ApiLicense
 import org.http4k.core.Uri
 import org.http4k.format.Argo
 
 class OpenApi3Test : ContractRendererContract<JsonNode>(Argo, OpenApi3(
-    ApiInfo("title", "1.2", "module description"),
+    ApiInfo("title", "1.2", "module description", "a short summary", ApiLicense.Apache2_0),
     Argo,
     listOf(AddSimpleFieldToRootNode),
     servers = listOf(ApiServer(Uri.of("https://localhost:8000")))