+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Given version `A.B.C.D`, breaking changes are to be expected in version number i
 ### v6.54.0.0 (uncut)
 - **http4k-***: Upgrade versions including Pug to v3.0.0
 - **http4k-template-pug4j**: [Breaking] Pug4J upgrade to v3.0.0 now uses a PugEngineBuilder instead of a PugConfiguration. The API is broadly the same (hopefully better!)
-- **http4k-api-openapi**: [Unlikely break] Updated to support OpenApi 3.2.0
+- **http4k-api-openapi**: [Unlikely break] Updated to support OpenApi 3.2.0, with various extra fields such as License and JSON Schema dialect.
 - **http4k-testing-webdriver-datastar**: [New module!] Datastar-enabled WebDriver implementation
 - **http4k-core**: [Fix] Handle UTF-8 percent-encoding for supplementary-plane characters in URI path segments
 - **http4k-connect-github**: [Fix] `HttpGitHub` will now properly authorize requests.  H/T @oharaandrew314